+++ calendar-service/src/main/scala/org/opentorah/calendar/ical/ICal.scala
@@ -1,8 +1,7 @@
 package org.opentorah.calendar.ical
 
-import org.opentorah.calendar.gregorian.Gregorian.Day
+import org.opentorah.calendar.roman.Gregorian.Day
 import java.io.PrintStream
-
 
 object ICal {
   type Properties = Seq[(String, String)]