+++ calendar/src/main/scala/org/opentorah/calendar/jewish/Shemittah.scala
@@ -1,6 +1,6 @@
 package org.opentorah.calendar.jewish
 
-import org.opentorah.dates.YearsCycle
+import org.opentorah.calendar.YearsCycle
 
 object Shemittah extends YearsCycle {
   override val first: Int = 1